refactor: refactor logic of strategy to handler
