+++ apps/govquests-frontend/src/components/Layout.tsx
@@ -5,7 +5,8 @@ import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { ConnectKitProvider, SIWEProvider } from "connectkit";
 import { config, siweConfig } from "@/wagmi";
 import Header from "./Header";
-import NotificationToaster from "@/domains/notifications/components/NotificationToaster";
+import { Toaster } from "./ui/toaster";
+import { useNotificationProcessor } from "@/domains/notifications/hooks/useNotificationProcessor";
 
 export const queryClient = new QueryClient();
 
@@ -29,10 +30,12 @@ export default function Layout({
 }: Readonly<{
   children: React.ReactNode;
 }>) {
+  useNotificationProcessor();
+
   return (
     <Providers>
       <div className="h-full">{children}</div>
-      <NotificationToaster />
+      <Toaster />
     </Providers>
   );
 }
