+++ apps/govquests-frontend/src/domains/notifications/strategies/QuestCompletedHandler.ts
@@ -1,0 +1,10 @@
+import { NotificationNode } from "../types/notificationTypes";
+import { OnNotificationHandler } from "./OnNotificationHandler";
+
+export class QuestCompletedHandler implements OnNotificationHandler {
+  constructor(private toast) {}
+
+  handle(notification: NotificationNode) {
+    this.toast({ title: notification.content });
+  }
+}
