--- apps/govquests-frontend/src/domains/notifications/strategies/QuestCompletedStrategy.ts
@@ -1,11 +1,0 @@
-import { NotificationStrategyProps } from "./NotificationStrategy";
-
-export const QuestCompletedStrategy = (props: NotificationStrategyProps) => {
-  const { notification, toast } = props;
-
-  const handleToast = () => {
-    toast({ title: notification.content });
-  };
-
-  return { handleToast };
-};
